refactor: composable for handling calendar events data
